# polaflixAlejandroAcebo/src/main/java/application/model/dto/seguimientoserie/SeguimientoSerieRequestDto.java
package application.model.dto.seguimientoserie;

import com.fasterxml.jackson.annotation.JsonProperty;

import jakarta.validation.constraints.Positive;
import lombok.AllArgsConstructor;
import lombok.Getter;

@Getter
@AllArgsConstructor
public class SeguimientoSerieRequestDto {
    
    @JsonProperty("idSerie")
    private int idSerie;
}