# swagger-json.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8081","description":"Generated server url"}],"paths":{"/usuarios/{id}":{"get":{"tags":["usuario-controller"],"operationId":"getUsuarioById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UsuarioResponseDto"}}}}}},"put":{"tags":["usuario-controller"],"operationId":"updateUsuario","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UsuarioUpdateDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UsuarioResponseDto"}}}}}},"delete":{"tags":["usuario-controller"],"operationId":"deleteUsuario","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK"}}}},"/series/{serieId}/temporadas/{temporadaId}/capitulos/{id}":{"get":{"tags":["capitulo-controller"],"operationId":"getCapituloById","parameters":[{"name":"serieId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"temporadaId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CapituloResponseDto"}}}}}},"put":{"tags":["capitulo-controller"],"operationId":"updateCapitulo","parameters":[{"name":"serieId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"temporadaId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CapituloRequestDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CapituloResponseDto"}}}}}},"delete":{"tags":["capitulo-controller"],"operationId":"deleteCapitulo","parameters":[{"name":"serieId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"temporadaId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK"}}}},"/series/{serieId}/temporadas/{id}":{"get":{"tags":["temporada-controller"],"operationId":"getTemporadaById","parameters":[{"name":"serieId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TemporadaResponseDto"}}}}}},"put":{"tags":["temporada-controller"],"operationId":"updateTemporada","parameters":[{"name":"serieId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TemporadaRequestDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TemporadaResponseDto"}}}}}},"delete":{"tags":["temporada-controller"],"operationId":"deleteTemporada","parameters":[{"name":"serieId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK"}}}},"/series/{id}":{"get":{"tags":["serie-controller"],"operationId":"getSerieById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SerieResponseDto"}}}}}},"put":{"tags":["serie-controller"],"operationId":"updateSerie","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SerieRequestDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SerieResponseDto"}}}}}},"delete":{"tags":["serie-controller"],"operationId":"deleteSerie","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK"}}}},"/usuarios":{"get":{"tags":["usuario-controller"],"operationId":"getUsuarios","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UsuarioResponseDto"}}}}}}},"post":{"tags":["usuario-controller"],"operationId":"createUsuario","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UsuarioRequestDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UsuarioResponseDto"}}}}}}},"/usuarios/{usuarioId}/visualizaciones":{"get":{"tags":["visualizacion-controller"],"operationId":"getVisualizaciones","parameters":[{"name":"usuarioId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/VisualizacionResponseDto"}}}}}}},"post":{"tags":["visualizacion-controller"],"operationId":"createVisualizacion","parameters":[{"name":"usuarioId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/VisualizacionRequestDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/VisualizacionResponseDto"}}}}}}},"/usuarios/{usuarioId}/seguimientos":{"get":{"tags":["seguimiento-serie-controller"],"operationId":"getSeguimientos","parameters":[{"name":"usuarioId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/SeguimientoSerieResponseDto"}}}}}}},"post":{"tags":["seguimiento-serie-controller"],"operationId":"createSeguimiento","parameters":[{"name":"usuarioId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SeguimientoSerieRequestDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SeguimientoSerieResponseDto"}}}}}}},"/series":{"get":{"tags":["serie-controller"],"operationId":"getSeries","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/SerieResponseDto"}}}}}}},"post":{"tags":["serie-controller"],"operationId":"createSerie","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SerieRequestDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SerieResponseDto"}}}}}}},"/series/{serieId}/temporadas":{"get":{"tags":["temporada-controller"],"operationId":"getTemporadas","parameters":[{"name":"serieId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TemporadaResponseDto"}}}}}}},"post":{"tags":["temporada-controller"],"operationId":"createTemporada","parameters":[{"name":"serieId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TemporadaRequestDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TemporadaResponseDto"}}}}}}},"/series/{serieId}/temporadas/{temporadaId}/capitulos":{"get":{"tags":["capitulo-controller"],"operationId":"getCapitulos","parameters":[{"name":"serieId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"temporadaId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CapituloResponseDto"}}}}}}},"post":{"tags":["capitulo-controller"],"operationId":"createCapitulo","parameters":[{"name":"serieId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"temporadaId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CapituloRequestDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CapituloResponseDto"}}}}}}},"/usuarios/{usuarioId}/visualizaciones/{id}":{"get":{"tags":["visualizacion-controller"],"operationId":"getVisualizacionById","parameters":[{"name":"usuarioId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/VisualizacionResponseDto"}}}}}},"delete":{"tags":["visualizacion-controller"],"operationId":"deleteVisualizacion","parameters":[{"name":"usuarioId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK"}}}},"/usuarios/{usuarioId}/seguimientos/{id}":{"get":{"tags":["seguimiento-serie-controller"],"operationId":"getSeguimientoById","parameters":[{"name":"usuarioId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SeguimientoSerieResponseDto"}}}}}},"delete":{"tags":["seguimiento-serie-controller"],"operationId":"deleteSeguimiento","parameters":[{"name":"usuarioId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK"}}}},"/planes":{"get":{"tags":["plan-controller"],"operationId":"getPlanes","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PlanResponseDto"}}}}}}}},"/planes/{id}":{"get":{"tags":["plan-controller"],"operationId":"getPlanById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PlanResponseDto"}}}}}}},"/facturas":{"get":{"tags":["factura-controller"],"operationId":"getFacturas","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/FacturaResponseDto"}}}}}}}},"/facturas/{id}":{"get":{"tags":["factura-controller"],"operationId":"getFacturaById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/FacturaResponseDto"}}}}}}},"/facturas/usuario/{usuarioId}":{"get":{"tags":["factura-controller"],"operationId":"getFacturasByUsuario","parameters":[{"name":"usuarioId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/FacturaResponseDto"}}}}}}}},"/usuarios/{usuarioId}/seguimientos/series/{serieId}":{"delete":{"tags":["seguimiento-serie-controller"],"operationId":"deleteSeguimientoBySerie","parameters":[{"name":"usuarioId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"serieId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK"}}}}},"components":{"schemas":{"UsuarioUpdateDto":{"type":"object","properties":{"nombre":{"type":"string"},"password":{"type":"string"},"cuentaBancaria":{"type":"string"},"idPlan":{"type":"integer","format":"int32"}}},"UsuarioResponseDto":{"type":"object","properties":{"idUsuario":{"type":"integer","format":"int32"},"nombre":{"type":"string"},"cuentaBancaria":{"type":"string"},"idPlan":{"type":"integer","format":"int32"}}},"CapituloRequestDto":{"required":["enlace","nombreCapitulo"],"type":"object","properties":{"idTemporada":{"type":"integer","format":"int32"},"nombreCapitulo":{"type":"string"},"numeroCapitulo":{"type":"integer","format":"int32"},"enlace":{"type":"string"},"descripcion":{"type":"string"}}},"CapituloResponseDto":{"type":"object","properties":{"idCapitulo":{"type":"integer","format":"int32"},"idTemporada":{"type":"integer","format":"int32"},"nombreCapitulo":{"type":"string"},"numeroCapitulo":{"type":"integer","format":"int32"},"enlace":{"type":"string"},"descripcion":{"type":"string"}}},"TemporadaRequestDto":{"required":["nombreTemporada"],"type":"object","properties":{"idSerie":{"type":"integer","format":"int32"},"nombreTemporada":{"type":"string"},"numeroTemporada":{"type":"integer","format":"int32"}}},"TemporadaResponseDto":{"type":"object","properties":{"idTemporada":{"type":"integer","format":"int32"},"idSerie":{"type":"integer","format":"int32"},"nombreTemporada":{"type":"string"},"numeroTemporada":{"type":"integer","format":"int32"}}},"SerieRequestDto":{"required":["nombreSerie"],"type":"object","properties":{"nombreSerie":{"type":"string"},"sinopsis":{"type":"string"},"tipoSerie":{"type":"string"}}},"SerieResponseDto":{"type":"object","properties":{"idSerie":{"type":"integer","format":"int32"},"nombreSerie":{"type":"string"},"sinopsis":{"type":"string"},"tipoSerie":{"type":"string"}}},"UsuarioRequestDto":{"required":["cuentaBancaria","idPlan","nombre","password"],"type":"object","properties":{"nombre":{"type":"string"},"password":{"type":"string"},"cuentaBancaria":{"type":"string"},"idPlan":{"type":"integer","format":"int32"}}},"VisualizacionRequestDto":{"type":"object","properties":{"idCapitulo":{"type":"integer","format":"int32"}}},"VisualizacionResponseDto":{"type":"object","properties":{"idVisualizacion":{"type":"integer","format":"int32"},"idUsuario":{"type":"integer","format":"int32"},"idCapitulo":{"type":"integer","format":"int32"},"fechaVisualizacion":{"type":"string"}}},"SeguimientoSerieRequestDto":{"type":"object","properties":{"idSerie":{"type":"integer","format":"int32"}}},"SeguimientoSerieResponseDto":{"type":"object","properties":{"idSeguimientoSerie":{"type":"integer","format":"int32"},"idSerie":{"type":"integer","format":"int32"},"idCapituloUltimoVisto":{"type":"integer","format":"int32"},"estadoSerie":{"type":"string"}}},"PlanResponseDto":{"type":"object","properties":{"idPlan":{"type":"integer","format":"int32"},"precio":{"type":"number","format":"double"},"tipoPlan":{"type":"string"}}},"FacturaResponseDto":{"type":"object","properties":{"idFactura":{"type":"integer","format":"int32"},"idUsuario":{"type":"integer","format":"int32"},"fecha":{"type":"string"},"total":{"type":"number","format":"double"}}}}}}